backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=1.4
alembic>=1.10.0
httpx>=0.24.0
openai>=1.30.0
pytest>=7.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
mcp>=1.0.0
